Refactor internal GCP module (#111)
* Initial refactor proposal (#93)


* Switch import from classes to modules (#96)

* Switch import from classes to modules


* Switch from aws/gcp_internal to internal


* Fix identation


* Fix most cyclic dependencies

* More import fix



* Remove forensics interface (#100)


* Remove duplicate code (#101)

* Initial refactor proposal (#93)


* Remove forensics interface (#100)


* Remove duplicate code


* refactoring gcp
