+++ libcloudforensics/providers/gcp/internal/compute.py
@@ -14,472 +14,460 @@
 # limitations under the License.
 """Google compute functionality."""
 
-import datetime
-import os
-import re
-import subprocess
-import time
+from googleapiclient.errors import HttpError
 
-from libcloudforensics.providers.gcp.internal.common import REGEX_DISK_NAME, LOGGER  # pylint: disable=line-too-long
-from libcloudforensics.providers.gcp.internal import project as gcp_project
+from libcloudforensics.providers.gcp.internal import common
+from libcloudforensics.providers.gcp.internal import compute_resources # pylint: disable=cyclic-import
+from libcloudforensics.scripts import utils
 
-
-class GoogleComputeBaseResource:
-  """Base class representing a Computer Engine resource.
-
+class GoogleCloudCompute(common.GoogleCloudComputeClient):
+  """Class representing all Google Cloud Compute objects in a project.
   Attributes:
-    project (GoogleCloudProject): Cloud project for the resource.
-    zone (str): What zone the resource is in.
-    name (str): Name of the resource.
-    labels (dict): Dictionary of labels for the resource, if existing.
+    project_id: Project name.
+    default_zone: Default zone to create new resources in.
   """
 
-  def __init__(self, project, zone, name, labels=None):
-    """Initialize the Google Compute Resource base object.
+  def __init__(self, project_id, default_zone=None):
+    """Initialize the Google Compute Resources in a project.
 
     Args:
-      project (GoogleCloudProject): Cloud project for the resource.
-      zone (str): What zone the resource is in.
-      name (str): Name of the resource.
-      labels (dict): Dictionary of labels for the resource, if existing.
+      project_id (str): Google Cloud project ID.
+      default_zone (str): Optional. Default zone to create new resources in.
+          None means GlobalZone.
     """
 
-    self.project = project
-    self.zone = zone
-    self.name = name
-    self.labels = labels
-    self._data = None
+    self.project_id = project_id
+    self.default_zone = default_zone
+    self._instances = None
+    self._disks = None
+    super(GoogleCloudCompute, self).__init__(self.project_id)
 
-  def GetValue(self, key):
-    """Get specific value from the resource key value store.
+  def Instances(self, refresh=True):
+    """Get all instances in the project.
 
     Args:
-      key (str): A key of type String to get key's corresponding value.
-
+      refresh (boolean): Optional. Returns refreshed result if True.
     Returns:
-      str: Value of key or None if key is missing.
+      dict: Dictionary mapping instance names (str) to their respective
+          GoogleComputeInstance object.
     """
+    if not(refresh) and self._instances:
+      return self._instances
+    self._instances = self.ListInstances()
+    return self._instances
 
-    self._data = self.GetOperation()  # pylint: disable=no-member
-    return self._data.get(key)
-
-  def GetSourceString(self):
-    """API URL to the resource.
-
-    Returns:
-      str: The full API URL to the resource.
-    """
-
-    if self._data:
-      return self._data['selfLink']
-    return self.GetValue('selfLink')
-
-  def GetResourceType(self):
-    """Get the resource type from the resource key-value store.
-
-    Returns:
-      str: Resource Type which is a string with one of the following values:
-          compute#instance
-          compute#disk
-          compute#Snapshot
-    """
-
-    if self._data:
-      return self._data['kind']
-    return self.GetValue('kind')
-
-  def FormOperation(self, operation_name):
-    """Form an API operation object for the compute resource.
-
-    Example:[RESOURCE].FormOperation('setLabels')(**kwargs)
-    [RESOURCE] can be type "instance", disk or "Snapshot".
+  def Disks(self, refresh=True):
+    """Get all instances in the project.
 
     Args:
-      operation_name (str): The name of the API operation you need to perform.
+      refresh (boolean): Optional. Returns refreshed result if True.
+    Returns:
+      dict: Dictionary mapping instance names (str) to their respective
+          GoogleComputeInstance object.
+    """
+    if not(refresh) and self._instances:
+      return self._instances
+    self._instances = self.ListInstances()
+    return self._instances
+
+  def ListInstances(self):
+    """List instances in project.
 
     Returns:
-      apiclient.discovery.Resource: An API operation object for the
-          referenced compute resource.
+      dict: Dictionary mapping instance names (str) to their respective
+          GoogleComputeInstance object.
+    """
+
+    have_all_tokens = False
+    page_token = None
+    instances = {}
+    while not have_all_tokens:
+      gce_instance_client = self.GceApi().instances()
+      if page_token:
+        request = gce_instance_client.aggregatedList(
+            project=self.project_id, pageToken=page_token)
+      else:
+        request = gce_instance_client.aggregatedList(project=self.project_id)
+      result = request.execute()
+      page_token = result.get('nextPageToken')
+      if not page_token:
+        have_all_tokens = True
+
+      for zone in result['items']:
+        try:
+          for instance in result['items'][zone]['instances']:
+            _, zone = instance['zone'].rsplit('/', 1)
+            name = instance['name']
+            instances[name] = compute_resources.GoogleComputeInstance(
+                self.project_id, zone, name, labels=instance.get('labels'))
+        except KeyError:
+          pass
+
+    return instances
+
+  def ListDisks(self):
+    """List disks in project.
+
+    Returns:
+      dict: Dictionary mapping disk names (str) to their respective
+          GoogleComputeDisk object.
+    """
+
+    have_all_tokens = False
+    page_token = None
+    disks = {}
+    while not have_all_tokens:
+      gce_disk_client = self.GceApi().disks()
+      if page_token:
+        request = gce_disk_client.aggregatedList(
+            project=self.project_id, pageToken=page_token)
+      else:
+        request = gce_disk_client.aggregatedList(project=self.project_id)
+      result = request.execute()
+      page_token = result.get('nextPageToken')
+      if not page_token:
+        have_all_tokens = True
+      for zone in result['items']:
+        try:
+          for disk in result['items'][zone]['disks']:
+            _, zone = disk['zone'].rsplit('/', 1)
+            name = disk['name']
+            disks[name] = compute_resources.GoogleComputeDisk(
+                self.project_id, zone, name, labels=disk.get('labels'))
+        except KeyError:
+          pass
+
+    return disks
+
+  def GetInstance(self, instance_name):
+    """Get instance from project.
+
+    Args:
+      instance_name (str): The instance name.
+
+    Returns:
+      GoogleComputeInstance: A Google Compute Instance object.
 
     Raises:
-      RuntimeError: If resource type is not defined as a type which
-          extends the GoogleComputeBaseResource class.
+      RuntimeError: If instance does not exist.
     """
 
-    resource_type = self.GetResourceType()
-    module = None
-    if resource_type not in ['compute#instance', 'compute#Snapshot',
-                             'compute#disk']:
-      error_msg = (
-          'Compute resource Type {0:s} is not one of the defined '
-          'types in libcloudforensics library '
-          '(Instance, Disk or Snapshot).').format(resource_type)
+    instances = self.Instances()
+    instance = instances.get(instance_name)
+    if not instance:
+      error_msg = 'Instance {0:s} was not found in project {1:s}'.format(
+          instance_name, self.project_id)
       raise RuntimeError(error_msg)
-    if resource_type == 'compute#instance':
-      module = self.project.GceApi().instances()
-    elif resource_type == 'compute#disk':
-      module = self.project.GceApi().disks()
-    elif resource_type == 'compute#Snapshot':
-      module = self.project.GceApi().snapshots()
-
-    operation_func_to_call = getattr(module, operation_name)
-    return operation_func_to_call
-
-  def GetLabels(self):
-    """Get all labels of a compute resource.
-
-    Returns:
-      dict: A dictionary of all labels.
-    """
-
-    operation = self.GetOperation()  # pylint: disable=no-member
-
-    return operation.get('labels')
-
-  def AddLabels(self, new_labels_dict, blocking_call=False):
-    """Add or update labels of a compute resource.
-
-    Args:
-      new_labels_dict (dict): A dictionary containing the labels to be added,
-          ex:{"incident_id": "1234abcd"}.
-      blocking_call (bool): Optional. A boolean to decide whether the API call
-          should be blocking or not, default is False.
-
-    Returns:
-      dict: The response of the API operation.
-
-    Raises:
-      RuntimeError: If the Compute resource Type is not one of instance,
-          disk or snapshot.
-    """
-
-    get_operation = self.GetOperation()  # pylint: disable=no-member
-    label_fingerprint = get_operation['labelFingerprint']
-
-    existing_labels_dict = {}
-    if self.GetLabels() is not None:
-      existing_labels_dict = self.GetLabels()
-    existing_labels_dict.update(new_labels_dict)
-    labels_dict = existing_labels_dict
-    request_body = {
-        'labels': labels_dict,
-        'labelFingerprint': label_fingerprint
-    }
-
-    resource_type = self.GetResourceType()
-    response = None
-    if resource_type not in ['compute#instance', 'compute#Snapshot',
-                             'compute#disk']:
-      error_msg = (
-          'Compute resource Type {0:s} is not one of the defined '
-          'types in libcloudforensics library '
-          '(Instance, Disk or Snapshot) ').format(resource_type)
-      raise RuntimeError(error_msg)
-    if resource_type == 'compute#instance':
-      response = self.FormOperation('setLabels')(
-          instance=self.name, project=self.project.project_id, zone=self.zone,
-          body=request_body).execute()
-    elif resource_type == 'compute#disk':
-      response = self.FormOperation('setLabels')(
-          resource=self.name, project=self.project.project_id, zone=self.zone,
-          body=request_body).execute()
-    elif resource_type == 'compute#Snapshot':
-      response = self.FormOperation('setLabels')(
-          resource=self.name, project=self.project.project_id,
-          body=request_body).execute()
-    if blocking_call:
-      self.project.BlockOperation(response, zone=self.zone)
-
-    return response
-
-
-class GoogleComputeInstance(GoogleComputeBaseResource):
-  """Class representing a Google Compute Engine virtual machine."""
-
-  def GetOperation(self):
-    """Get API operation object for the virtual machine.
-
-    Returns:
-      dict: An API operation object for a Google Compute Engine
-          virtual machine.
-    """
-
-    gce_instance_client = self.project.GceApi().instances()
-    request = gce_instance_client.get(
-        instance=self.name, project=self.project.project_id, zone=self.zone)
-    response = request.execute()
-    return response
-
-  def GetBootDisk(self):
-    """Get the virtual machine boot disk.
-
-    Returns:
-      GoogleComputeDisk: Disk object or None if no disk can be found.
-    """
-
-    for disk in self.GetValue('disks'):
-      if disk['boot']:
-        disk_name = disk['source'].split('/')[-1]
-        return self.project.GetDisk(disk_name=disk_name)
-    return None
+    return instance
 
   def GetDisk(self, disk_name):
-    """Gets a disk attached to this virtual machine disk by name.
+    """Get a GCP disk object.
 
     Args:
-      disk_name (str): The name of the disk to get.
+      disk_name (str): Name of the disk.
 
     Returns:
       GoogleComputeDisk: Disk object.
 
     Raises:
-      RuntimeError: If disk name is not found among those attached to the
-          instance.
+      RuntimeError: When the specified disk cannot be found in project.
     """
 
-    for disk in self.GetValue('disks'):
-      if disk['source'].split('/')[-1] == disk_name:
-        return self.project.GetDisk(disk_name=disk_name)
-    error_msg = 'Disk name "{0:s}" not attached to instance'.format(disk_name)
-    raise RuntimeError(error_msg)
+    disks = self.Disks()
+    disk = disks.get(disk_name)
+    if not disk:
+      error_msg = 'Disk {0:s} was not found in project {1:s}'.format(
+          disk_name, self.project_id)
+      raise RuntimeError(error_msg)
+    return disk
 
-  def ListDisks(self):
-    """List all disks for the virtual machine.
-
-    Returns:
-      list(str): List of disk names.
-    """
-
-    return [disk['source'].split('/')[-1] for disk in self.GetValue('disks')]
-
-  def __SshConnection(self):
-    """Create an SSH connection to the virtual machine."""
-
-    devnull = open(os.devnull, 'w')
-    subprocess.check_call([
-        'gcloud', 'compute', '--project', self.project.project_id, 'ssh',
-        '--zone', self.zone, self.name
-    ], stderr=devnull)
-
-  def Ssh(self):
-    """Connect to the virtual machine over SSH."""
-
-    max_retries = 100  # times to retry the connection
-    retries = 0
-
-    LOGGER.info(
-        self.project.FormatLogMessage('Connecting to analysis VM over SSH'))
-
-    while retries < max_retries:
-      try:
-        self.__SshConnection()
-        break
-      except subprocess.CalledProcessError:
-        retries += 1
-        time.sleep(5)  # seconds between connections
-
-  def AttachDisk(self, disk, read_write=False):
-    """Attach a disk to the virtual machine.
+  def CreateDiskFromSnapshot(  # pylint: disable=missing-type-doc
+      self, snapshot, disk_name=None, disk_name_prefix='',
+      disk_type='pd-standard'):
+    """Create a new disk based on a Snapshot.
 
     Args:
-      disk (GoogleComputeDisk): Disk to attach.
-      read_write (bool): Optional. Boolean indicating whether the disk should
-          be attached in RW mode. Default is False (read-only).
+      snapshot (libcloudforensics.providers.gcp.internal.compute_resources
+          .GoogleComputeSnapshot): Snapshot to use.
+      disk_name (str): Optional. String to use as new disk name.
+      disk_name_prefix (str): Optional. String to prefix the disk name with.
+      disk_type (str): Optional. URL of the disk type resource describing
+          which disk type to use to create the disk. Default is pd-standard. Use
+          pd-ssd to have a SSD disk. You can list all available disk types by
+          running the following command: gcloud compute disk-types list
+
+    Returns:
+      GoogleComputeDisk: Google Compute Disk.
+
+    Raises:
+      RuntimeError: If the disk exists already.
     """
 
-    mode = 'READ_ONLY'  # Default mode
-    if read_write:
-      mode = 'READ_WRITE'
-
-    LOGGER.info(
-        self.project.FormatLogMessage(
-            'Attaching {0} to VM {1} in {2} mode'.format(
-                disk.name, self.name, mode)))
-
-    operation_config = {
-        'mode': mode,
-        'source': disk.GetSourceString(),
-        'boot': False,
-        'autoDelete': False,
+    if not disk_name:
+      disk_name = common.GenerateDiskName(snapshot, disk_name_prefix)
+    body = {
+        'name':
+            disk_name,
+        'sourceSnapshot':
+            snapshot.GetSourceString(),
+        'type':
+            'projects/{0:s}/zones/{1:s}/diskTypes/{2:s}'.format(
+                self.project_id, self.default_zone, disk_type)
     }
-    gce_instance_client = self.project.GceApi().instances()
-    request = gce_instance_client.attachDisk(
-        instance=self.name, project=self.project.project_id, zone=self.zone,
-        body=operation_config)
-    response = request.execute()
-    self.project.BlockOperation(response, zone=self.zone)
+    try:
+      gce_disks_client = self.GceApi().disks()
+      request = gce_disks_client.insert(
+          project=self.project_id, zone=self.default_zone, body=body)
+      response = request.execute()
+    except HttpError as exception:
+      if exception.resp.status == 409:
+        error_msg = 'Disk {0:s} already exists'.format(disk_name)
+        raise RuntimeError(error_msg)
+      error_msg = (
+          'Unknown error (status: {0:d}) occurred when creating disk '
+          'from Snapshot:\n{1!s}').format(exception.resp.status, exception)
+      raise RuntimeError(error_msg)
+    self.BlockOperation(response, zone=self.default_zone)
+    return compute_resources.GoogleComputeDisk(
+        project_id=self.project_id, zone=self.default_zone, name=disk_name)
 
+  def GetOrCreateAnalysisVm(self,
+                            vm_name,
+                            boot_disk_size,
+                            disk_type='pd-standard',
+                            cpu_cores=4,
+                            image_project='ubuntu-os-cloud',
+                            image_family='ubuntu-1804-lts',
+                            packages=None):
+    """Get or create a new virtual machine for analysis purposes.
 
-class GoogleComputeDisk(GoogleComputeBaseResource):
-  """Class representing a Compute Engine disk."""
-
-  def GetOperation(self):
-    """Get API operation object for the disk.
-
-    Returns:
-      dict: An API operation object for a Google Compute Engine disk.
-    """
-
-    gce_disk_client = self.project.GceApi().disks()
-    request = gce_disk_client.get(
-        disk=self.name, project=self.project.project_id, zone=self.zone)
-    response = request.execute()
-    return response
-
-  def Snapshot(self, snapshot_name=None):
-    """Create Snapshot of the disk.
-
-    The Snapshot name must comply with the following RegEx:
-      - ^(?=.{1,63}$)[a-z]([-a-z0-9]*[a-z0-9])?$
-
-    i.e., it must be between 1 and 63 chars, the first character must be a
-    lowercase letter, and all following characters must be a dash, lowercase
-    letter, or digit, except the last character, which cannot be a dash.
+    If none of the optional parameters are specified, then by default the
+    analysis VM that will be created will run Ubuntu 18.04 LTS. A default
+    set of forensic tools is also installed (a custom one may be provided
+    using the 'packages' argument).
 
     Args:
-      snapshot_name (str): Optional. Name of the Snapshot.
+      vm_name (str): Name of the virtual machine.
+      boot_disk_size (int): The size of the analysis VM boot disk (in GB).
+      disk_type (str): Optional. URL of the disk type resource describing
+          which disk type to use to create the disk. Default is pd-standard. Use
+          pd-ssd to have a SSD disk.
+      cpu_cores (int): Optional. Number of CPU cores for the virtual machine.
+      image_project (str): Optional. Name of the project where the analysis VM
+          image is hosted.
+      image_family (str): Optional. Name of the image to use to create the
+          analysis VM.
+      packages (list[str]): Optional. List of packages to install in the VM.
 
     Returns:
-      GoogleComputeSnapshot: A Snapshot object.
+      tuple(GoogleComputeInstance, bool): A tuple with a virtual machine object
+          and a boolean indicating if the virtual machine was created or not.
 
     Raises:
-      ValueError: If the name of the snapshot does not comply with the RegEx.
+      RuntimeError: If virtual machine cannot be created.
     """
 
-    timestamp = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
-    if not snapshot_name:
-      snapshot_name = self.name
-    truncate_at = 63 - len(timestamp) - 1
-    snapshot_name = '{0}-{1}'.format(snapshot_name[:truncate_at], timestamp)
-    if not REGEX_DISK_NAME.match(snapshot_name):
-      raise ValueError(
-          'Snapshot name {0:s} does not comply with '
-          '{1:s}'.format(snapshot_name, REGEX_DISK_NAME.pattern))
-    LOGGER.info(
-        self.project.FormatLogMessage(
-            'New Snapshot: {0}'.format(snapshot_name)))
-    operation_config = {'name': snapshot_name}
-    gce_disk_client = self.project.GceApi().disks()
-    request = gce_disk_client.createSnapshot(
-        disk=self.name, project=self.project.project_id, zone=self.zone,
-        body=operation_config)
-    response = request.execute()
-    self.project.BlockOperation(response, zone=self.zone)
-    return GoogleComputeSnapshot(disk=self, name=snapshot_name)
+    if not self.default_zone:
+      raise RuntimeError('Cannot create VM, zone information is missing')
 
+    # Re-use instance if it already exists, or create a new one.
+    try:
+      instance = self.GetInstance(vm_name)
+      created = False
+      return instance, created
+    except RuntimeError:
+      pass
 
-class GoogleComputeSnapshot(GoogleComputeBaseResource):
-  """Class representing a Compute Engine Snapshot.
+    machine_type = 'zones/{0}/machineTypes/n1-standard-{1:d}'.format(
+        self.default_zone, cpu_cores)
+    ubuntu_image = self.GceApi().images().getFromFamily(
+        project=image_project, family=image_family).execute()
+    source_disk_image = ubuntu_image['selfLink']
 
-  Attributes:
-    disk (GoogleComputeDisk): Disk used for the Snapshot.
-  """
+    startup_script = utils.ReadStartupScript()
 
-  def __init__(self, disk, name):
-    """Initialize the Snapshot object.
+    if packages:
+      startup_script = startup_script.replace(
+          '${packages[@]}', ' '.join(packages))
 
-    Args:
-      disk (GoogleComputeDisk): Disk used for the Snapshot.
-      name (str): Name of the Snapshot.
-    """
-
-    super(GoogleComputeSnapshot, self).__init__(
-        project=disk.project, zone=None, name=name)
-    self.disk = disk
-
-  def GetOperation(self):
-    """Get API operation object for the Snapshot.
-
-    Returns:
-      dict: An API operation object for a Google Compute Engine Snapshot.
-    """
-
-    gce_snapshot_client = self.project.GceApi().snapshots()
-    request = gce_snapshot_client.get(
-        snapshot=self.name, project=self.project.project_id)
-    response = request.execute()
-    return response
-
-  def Delete(self):
-    """Delete a Snapshot."""
-
-    LOGGER.info(
-        self.project.FormatLogMessage(
-            'Deleted Snapshot: {0}'.format(self.name)))
-    gce_snapshot_client = self.project.GceApi().snapshots()
-    request = gce_snapshot_client.delete(
-        project=self.project.project_id, snapshot=self.name)
-    response = request.execute()
-    self.project.BlockOperation(response)
-
-
-class GoogleComputeImage(GoogleComputeBaseResource):
-  """Class representing a Compute Engine Image.
-
-  Attributes:
-    disk (GoogleComputeDisk): Disk used for the Snapshot.
-  """
-
-  def GetOperation(self):
-    """Get API operation object for the image.
-
-    Returns:
-      dict: Holding an API operation object for a Google Compute Engine Image.
-    """
-
-    gce_image_client = self.project.GceApi().images()
-    request = gce_image_client.get(
-        project=self.project.project_id, image=self.name)
-    response = request.execute()
-    return response
-
-  def ExportImage(self, gcs_output_folder, output_name=None):
-    """Export compute image to Google Cloud storage.
-
-    Exported image is compressed and stored in .tar.gz format.
-
-    Args:
-      gcs_output_folder (str): Folder path of the exported image.
-      output_name (str): Optional. Name of the output file, must end with
-          .tar.gz, if not exist, the [image_name].tar.gz will be used.
-
-    Raises:
-      RuntimeError: If exported image name is invalid.
-    """
-
-    if output_name:
-      if not bool(re.match("^[A-Za-z0-9-]*$", output_name)):
-        raise RuntimeError(
-            'Destination disk name must comply with expression ^[A-Za-z0-9-]*$')
-      full_path = '{0:s}.tar.gz'.format(
-          os.path.join(gcs_output_folder, output_name))
-    else:
-      full_path = '{0:s}.tar.gz'.format(
-          os.path.join(gcs_output_folder, self.name))
-    build_body = {
-        'timeout': '86400s',
-        'steps': [{
-            'args': [
-                '-source_image={0:s}'.format(self.name),
-                '-destination_uri={0:s}'.format(full_path),
-                '-client_id=api',
-            ],
-            'name': 'gcr.io/compute-image-tools/gce_vm_image_export:release',
-            'env': []
+    config = {
+        'name': vm_name,
+        'machineType': machine_type,
+        'disks': [{
+            'boot': True,
+            'autoDelete': True,
+            'initializeParams': {
+                'diskType':
+                    'projects/{0:s}/zones/{1:s}/diskTypes/{2:s}'.format(
+                        self.project_id, self.default_zone, disk_type),
+                'sourceImage':
+                    source_disk_image,
+                'diskSizeGb':
+                    boot_disk_size,
+            }
         }],
-        'tags': ['gce-daisy', 'gce-daisy-image-export']
+        'networkInterfaces': [{
+            'network':
+                'global/networks/default',
+            'accessConfigs': [{
+                'type': 'ONE_TO_ONE_NAT',
+                'name': 'External NAT'
+            }]
+        }],
+        'serviceAccounts': [{
+            'email':
+                'default',
+            'scopes': [
+                'https://www.googleapis.com/auth/devstorage.read_write',
+                'https://www.googleapis.com/auth/logging.write'
+            ]
+        }],
+        'metadata': {
+            'items': [{
+                'key': 'startup-script',
+                # Analysis software to install.
+                'value': startup_script
+            }]
+        }
     }
-    cloud_build = gcp_project.GoogleCloudBuild(self.project.project_id)
-    response = cloud_build.CreateBuild(build_body)
-    cloud_build.BlockOperation(response)
-    LOGGER.info('Image {0:s} exported to {1:s}.'.format(self.name, full_path))
+    gce_instance_client = self.GceApi().instances()
+    request = gce_instance_client.insert(
+        project=self.project_id, zone=self.default_zone, body=config)
+    response = request.execute()
+    self.BlockOperation(response, zone=self.default_zone)
+    instance = compute_resources.GoogleComputeInstance(
+        project_id=self.project_id, zone=self.default_zone, name=vm_name)
+    created = True
+    return instance, created
 
-  def Delete(self):
-    """Delete Compute Disk Image from a project.
+  def ListInstanceByLabels(self, labels_filter, filter_union=True):
+    """List VMs in a project with one/all of the provided labels.
+
+    This will call the __ListByLabel on instances() API object
+    with the proper labels filter and return a dict with name and metadata
+    for each instance, e.g.:
+        {'instance-1': {'zone': 'us-central1-a', 'labels': {'id': '123'}}
+
+    Args:
+      labels_filter (dict): A dict of labels to find e.g. {'id': '123'}.
+      filter_union (bool): Optional. A Boolean; True to get the union of all
+          filters, False to get the intersection.
+
+    Returns:
+      dict: Dictionary mapping instances to their respective
+          GoogleComputeInstance object.
     """
 
-    gce_image_client = self.project.GceApi().images()
-    request = gce_image_client.delete(
-        project=self.project.project_id, image=self.name)
+    instance_service_object = self.GceApi().instances()
+    return self.__ListByLabel(
+        labels_filter, instance_service_object, filter_union)
+
+  def ListDiskByLabels(self, labels_filter, filter_union=True):
+    """List Disks in a project with one/all of the provided labels.
+
+    This will call the __ListByLabel on disks() API object
+    with the proper labels filter and return a dict with name and metadata
+    for each disk, e.g.:
+        {'disk-1': {'zone': 'us-central1-a', 'labels': {'id': '123'}}
+
+    Args:
+      labels_filter (dict): A dict of labels to find e.g. {'id': '123'}.
+      filter_union (bool): Optional. A Boolean; True to get the union of all
+          filters, False to get the intersection.
+
+    Returns:
+      dict: Dictionary mapping disks to their respective GoogleComputeDisk
+          object.
+    """
+
+    disk_service_object = self.GceApi().disks()
+    return self.__ListByLabel(labels_filter, disk_service_object, filter_union)
+
+  def __ListByLabel(self, labels_filter, service_object, filter_union):
+    """List Disks/VMs in a project with one/all of the provided labels.
+
+    Private method used to select different compute resources by labels.
+
+    Args:
+      labels_filter (dict):  A dict of labels to find e.g. {'id': '123'}.
+      service_object (apiclient.discovery.Resource): Google Compute Engine
+          (Disk | Instance) service object.
+      filter_union (bool): A boolean; True to get the union of all filters,
+          False to get the intersection.
+
+    Returns:
+      dict: Dictionary mapping instances/disks to their respective
+          GoogleComputeInstance/GoogleComputeDisk object.
+
+    Raises:
+      RuntimeError: If the operation doesn't complete on GCP.
+    """
+
+    if not isinstance(filter_union, bool):
+      error_msg = (
+          'filter_union parameter must be of Type boolean {0:s} is an '
+          'invalid argument.').format(filter_union)
+      raise RuntimeError(error_msg)
+
+    resource_dict = {}
+    filter_expression = ''
+    operation = 'AND' if filter_union else 'OR'
+    for key, value in labels_filter.items():
+      filter_expression += 'labels.{0:s}={1:s} {2:s} '.format(
+          key, value, operation)
+    filter_expression = filter_expression[:-(len(operation) + 1)]
+
+    request = service_object.aggregatedList(
+        project=self.project_id, filter=filter_expression)
+    while request is not None:
+      response = request.execute()
+
+      for item in response['items'].items():
+        region_or_zone_string, resource_scoped_list = item
+
+        if 'warning' not in resource_scoped_list.keys():
+          _, zone = region_or_zone_string.rsplit('/', 1)
+          # Only one of the following loops will execute since the method is
+          # called either with a service object Instances or Disks.
+          for resource in resource_scoped_list.get('instances', []):
+            name = resource['name']
+            resource_dict[name] = compute_resources.GoogleComputeInstance(
+                self.project_id, zone, name, labels=resource['labels'])
+
+          for resource in resource_scoped_list.get('disks', []):
+            name = resource['name']
+            resource_dict[name] = compute_resources.GoogleComputeDisk(
+                self.project_id, zone, name, labels=resource['labels'])
+
+      request = service_object.aggregatedList_next(
+          previous_request=request, previous_response=response)
+    return resource_dict
+
+  def CreateImageFromDisk(self, src_disk, name=None):
+    """Creates an image from a persistent disk.
+
+    Args:
+      src_disk (GoogleComputeDisk): Source disk for the image.
+      name (str): Optional. Name of the image to create. Default
+          is [src_disk]_image.
+
+    Returns:
+      libcloudforensics.providers.gcp.internal.compute.GoogleComputeImage:
+          A Google Compute Image object.
+    """
+    if not name:
+      name = '{0:s}_image'.format(src_disk.name)
+    image_body = {
+        'name':
+            name,
+        'sourceDisk':
+            'projects/{project_id}/zones/{zone}/disks/{src_disk}'.format(
+                project_id=src_disk.project_id, zone=src_disk.zone,
+                src_disk=src_disk.name)
+    }
+    gce_image_client = self.GceApi().images()
+    request = gce_image_client.insert(
+        project=self.project_id, body=image_body, forceCreate=True)
     response = request.execute()
-    self.project.BlockOperation(response)
+    self.BlockOperation(response)
+    return compute_resources.GoogleComputeImage(self.project_id, None, name)
